+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix for GPU `coo_spmm` when nnz == 0
     - Fix MinkowskiInterpolationGPU for invalid samples (issue #383)
 - gradcheck wrap func 1.9
+- Handle `dense` for an empty tensor (issue #384)
 
 ## [0.5.4]